combine dropdump and nodetrace into nettrace
'dropdump' is an eBPF based tool, which is used to monitor packet
dropping and the drop reason by tracing 'kfree_skb' tracepoint. To
use this tool, the kernel should have the 'skb drop reason' feature,
which is described here:

https://lwn.net/Articles/885729/

'nodetrace' is an eBPF based tool which aims to mark and trace
packet between nodes and find which node the packet is dropped by.

Signed-off-by: Menglong Dong <imagedong@tencent.com>