rewrite nettrace base on libbpf
The BCC-based nettrace is too unwieldy and inefficient. Especially,
some environment is not able to run BCC.

Therefore, rewrite it with C language bases on libbpf. The new
version has more feture, such as 'intellect mode', which is able to
diagnose network problem and gives advise.

However, we don't delete BCC-based nettrace, and move it to the 'legacy'
directory. Users can still use it with 'nettrace-legacy' command.