bpf: support to parse skb togather with socket
In some kernel function, such as __tcp_transmit_skb(), the header of the
skb is not ready, so we can't parse it. However, we can parse part of it.
For example, we can get seq/ack_seq from skb_cb(), and get the
saddr/daddr/sport/dport from the socket.

Signed-off-by: Menglong Dong <dongml2@chinatelecom.cn>